# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API - Joseph Vigil"
  description: "Used to manage the everyday functionalities of a mechanic shop"
  version: "1.0.0"
host: "127.0.0.1:5000" #During development my API is served on my local host (will change when we go to prod)
schemes:
  - "http"
consumes:
  - "application/json" #MY api receives json data
produces:
  - "application/json" #and produces json data
securityDefinitions: #Describing how to add token auth to the endpoints
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers:
    post:
      tags:
        - Customers
      summary: Create Customer
      description: "Create a new customer in the system"
      parameters:
        - in: body
          name: body
          description: "Customer object that needs to be added to the system"
          required: true
          schema:
            $ref: '#/components/schemas/CustomerInput'

      responses:
        '200':
          description: Customer created successfully
          schema:
            $ref: '#/components/schemas/Customer'

    get:
      tags:
        - Customers
      summary: Get all Customers
      description: "Retrieve a list of all customers in the system"

      responses:
        '200':
          description: List of all customers retrieved successfully
  /customers/{id}:
    get:
      tags:
        - Customers
      summary: Get Customer by ID
      description: "Retrieve a specific customer by their ID"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

    put:
      tags:
        - Customers
      summary: Update Customer
      description: "Update an existing customer in the system"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: body
          name: body
          description: "Customer object that needs to be added to the system"
          required: true
          schema:
            $ref: '#/components/schemas/CustomerInput'

      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

    delete:
      tags:
        - Customers
      summary: Delete Customer
      description: "Delete a customer from the system by their ID"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Customer deleted
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: Create Mechanic
      description: "Create a new mechanic in the system"
      parameters:
        - in: body
          name: body
          description: "Mechanic object that needs to be added to the system"
          required: true
          schema:
            $ref: '#/components/schemas/MechanicInput'
            
      responses:
        '200':
          description: Mechanic created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mechanic'

    get:
      tags:
        - Mechanics
      summary: Get all Mechanics
      description: "Retrieve a list of all mechanics in the system"

      responses:
        '200':
          description: List of all mechanics retrieved successfully
    
    put:
      tags:
        - Mechanics
      summary: Update Mechanic
      description: "Update an existing mechanic in the system"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: body
          name: body
          description: "Mechanic object that needs to be updated in the system"
          required: true
          schema:
            $ref: '#/components/schemas/MechanicInput'

      responses:
        '200':
          description: Mechanic updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mechanic'

    delete:
        tags:
          - Mechanics
        summary: Delete Mechanic
        description: "Delete a mechanic from the system by their ID"
        parameters:
          - in: path
            name: id
            required: true
            schema: { type: integer }
        responses:
          '200':
            description: Mechanic deleted
            content:
              application/json:
                schema:
                  type: object
                  properties: { message: { type: string } }
  /mechanics/{id}:
    get:
      tags:
        - Mechanics
      summary: Get Mechanic by ID
      description: "Retrieve a specific mechanic by their ID"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Mechanic found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mechanic'
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: Mechanic Login
      description: "Authenticate a mechanic and receive a bearer token for future requests"
      security:
       - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: "Login credentials for mechanic"
          required: true
          schema:
            type: object
            properties:
              email: { type: string }
              password: { type: string }

      responses:
        '200':
          description: Bearer Token
  /mechanics/profile:
    get:
      tags:
        - Mechanics
      summary: Get Mechanic Profile (requires token)
      description: "Retrieve the profile information of the authenticated mechanic"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Mechanic profile information
  /mechanics/my-tickets:
    get:
      tags:
        - Mechanics
      summary: Get My Tickets (requires token)
      description: "Retrieve a list of tickets assigned to the authenticated mechanic"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of assigned tickets
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CustomerInput:
      type: object
      properties:
        address: { type: string }
        email: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        phone: { type: string }
    Customer:
      allOf:
        - $ref: '#/components/schemas/CustomerInput'
        - type: object
          properties:
            id: { type: integer }
    MechanicInput:
      type: object
      properties:
        address: { type: string }
        email: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        password: { type: string }
        salary: { type: number }
    Mechanic:
      allOf:
        - $ref: '#/components/schemas/MechanicInput'
        - type: object
          properties:
            id: { type: integer }
...